\name{ggcounty-package}
\alias{ggcounty-package}
\alias{ggcounty}
\docType{package}
\title{
Generate ggplot2 geom_map county maps
}
\description{

}
\details{
\tabular{ll}{
Package: \tab ggcounty\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-16\cr
License: \tab MIT\cr
}
}
\author{
Bob Rudis (bob@rudis.net) @hrbrmstr

Maintainer: Who to complain to <bob@rudis.net>
}
\references{
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
