% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcounty.R
\name{ggcounty.us_full}
\alias{ggcounty.us_full}
\title{Returns a ggplot2 object with a geom_map of the lower 48 states and the non-contiguous states of Alaska and Hawaii}
\usage{
ggcounty.us_full(fill = "white", color = "#7f7f7f", size = 0.25,
  alpha = 1)
}
\arguments{
\item{fill}{color string (e.g. + default = "white")}

\item{border}{color string (e.g. + default = "#7f7f7f")}

\item{border}{line width (e.g. + default = 0.25)}

\item{fill}{alpha (e.g. + default = 1)}
}
\value{
list consisting of the fortified map object (map), list of FIPS ids (fips) & the ggplot2 object (gg)
}
\description{
Returns a ggplot2 object with a geom_map of the lower 48 states and the non-contiguous states of Alaska and Hawaii
}
\examples{
g <- ggcounty.us_full()
}

